requirements.txt
appdirs==1.4.3
attrs==19.3.0
backcall==0.1.0
black==19.10b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.1
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
numpy==1.17.4
pandas==0.25.3
parso==0.5.1
pathspec==0.6.0
pickleshare==0.7.5
pocket-api==0.1.5
prompt-toolkit==2.0.10
Pygments==2.4.2
python-dateutil==2.8.1
pytz==2019.3
regex==2019.11.1
requests==2.22.0
six==1.13.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.7
